Rename queries.ts to get-latest-subscription.ts
Rename the file to better reflect its single exported function.
Also update all import paths accordingly.
